Flat array constructor.
Fixes #27. Also improves Delaunay.from so that it more closely matches
Array.from: the specified points can be an array-like object with a length
property or an iterable; an optional this argument can be specified; and the x-
and y-accessors are invoked with the element, index and object as arguments.