The issue https://github.com/d3/d3-delaunay/issues/119 comes from the circumcenters we assign to the degenerate triangles created by points of the hull that are almost exactly aligned.
The "circumcenter" of such a triangle is a point at the infinite, in the direction orthogonal to the hull.

It must also be "away from the center". But we can't rely on the sign of *ab*, since it's itself a difference of two products and can be of the wrong sign, leading to a ray that sometimes goes in the opposite direction.

To overcome this problem, we use the first point of the first triangle as a guide to tell us which way is "away from the center".